+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for range (2XX) responses. [#1699](https://github.com/microsoft/kiota/issues/1699)
 - Added none output formatter to CLI commons. (Shell)
 - Added 'Accept' field of http request header in Ruby.  [#1660](https://github.com/microsoft/kiota/issues/1660)
 - Added support for text serialization in Python. [#1406](https://github.com/microsoft/kiota/issues/1406)